added simulation gui, moved Fit() to utility script

+++ Assets/Weapon.cs
@@ -4,9 +4,9 @@ using System.Collections;
 public class Weapon{
 	public Texture2D image;
 	public string name;
-	public int damage;
+	public float damage;
 	
-	public Weapon(Texture2D _image, string _name, int _damage){
+	public Weapon(Texture2D _image, string _name, float _damage){
 		image = _image;
 		name = _name;
 		damage = _damage;
